refactor mode messageStatusEventSchema into it's file
