+++ apps/etl-func/src/domain/message-status-event.ts
@@ -1,12 +1,29 @@
-/**
- * Transform a MessageStatus into a MessageStatusEvent
- **/
+import { timestampSchema } from "io-messages-common/types/date";
+import * as z from "zod";
 
 import {
   MessageStatus,
-  MessageStatusEvent,
-  messageStatusEventSchema,
+  messageStatusIdSchema,
+  statusEnum,
 } from "./message-status.js";
+
+export const messageStatusEventSchema = z.object({
+  created_at: z.number(),
+  id: messageStatusIdSchema,
+  is_archived: z.boolean(),
+  is_read: z.boolean(),
+  message_id: z.string().ulid(),
+  op: z.enum(["CREATE", "UPDATE", "DELETE"]),
+  schema_version: z.literal(1),
+  status: statusEnum,
+  timestamp: timestampSchema,
+  version: z.number().gte(0).int(),
+});
+export type MessageStatusEvent = z.TypeOf<typeof messageStatusEventSchema>;
+
+/**
+ * Transform a MessageStatus into a MessageStatusEvent
+ **/
 
 export const getMessageStatusEventFromMessageStatus = (
   messageStatus: MessageStatus,
